[CHORE] : GetServiceCombine 파일 이동
